Move remaining functions from glue.clients.util to glue.core.util and removed glue.clients
